+++ Pages/Admin/Precios.cshtml
@@ -1,6 +1,6 @@
 ﻿@page
 @using EstimadorBM.Models;
-@model EstimadorBM.Pages.Presupuesto.PreciosModel
+@model EstimadorBM.Pages.Admin.PreciosModel
 @{
 	ViewData["Title"] = "Presupuestos 73 Precios";
 }
