+++ Pages/Shared/_Layout.cshtml
@@ -18,6 +18,7 @@
 	<script src="https://cdn.jsdelivr.net/npm/bootstrap-dark-5@1.1.3/dist/js/darkmode.min.js"></script>
 	<link rel="stylesheet" href="https://cdn.jsdelivr.net/npm/bootstrap-icons@1.11.1/font/bootstrap-icons.css">
 
+	<base href="~/" />
 </head>
 <body class="d-flex flex-column h-100">
 	<header class="custom">
@@ -49,7 +50,6 @@
 	<main role="main" class="flex-shrink-0 pb-3">
 		<div class="container">
 			@RenderBody()
-
 		</div>
 	</main>
